+++ test/test_helper.rb
@@ -5,5 +5,6 @@ require "minitest/autorun"
 require "minitest/pride"
 require "csv"
 require "daru"
+require "rover"
 
 require_relative "support/active_record"