+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.3 (unreleased)
+
+- Added support for Rover
+- Raise error when missing user or item ids
+- Fixed string keys for Daru data frames
+
 ## 0.1.2 (2020-03-26)
 
 - Added experimental `optimize_item_recs` and `optimize_similar_users` methods